test(ui): split model-override lifecycle coverage into a focused file (max-lines)
